Reorganised file structure to make more sense (and allow easier future additions)
