+++ src/app/algorithm-page/algorithms/gale-shapley/gale-shapley.interface.ts
