+++ src/app/algorithm-page/algorithms/gale-shapley/gale-shapley.service.ts
