+++ src/app/algorithm-page/algorithms/simple/simple.interface.ts
