+++ src/app/algorithm-page/algorithms/simple/simple.service.spec.ts
