+++ src/app/algorithm-page/mat-animated-icon/mat-animated-icon.component.spec.ts
